fix: pass string enum literal into function that take string
