+++ packages/eslint-plugin/src/rules/semi.ts
@@ -2,6 +2,24 @@ import { TSESTree, TSESLint, AST_NODE_TYPES } from '@typescript-eslint/utils';
 import { getESLintCoreRule } from '../util/getESLintCoreRule';
 import * as util from '../util';
 
+/**
+ * The following nodes are handled by the `member-delimiter-style` rule:
+ * - AST_NODE_TYPES.TSCallSignatureDeclaration
+ * - AST_NODE_TYPES.TSConstructSignatureDeclaration
+ * - AST_NODE_TYPES.TSIndexSignature
+ * - AST_NODE_TYPES.TSMethodSignature
+ * - AST_NODE_TYPES.TSPropertySignature
+ */
+const SEMI_AST_NODE_TYPES = [
+  AST_NODE_TYPES.PropertyDefinition,
+  AST_NODE_TYPES.TSAbstractPropertyDefinition,
+  AST_NODE_TYPES.TSDeclareFunction,
+  AST_NODE_TYPES.TSExportAssignment,
+  AST_NODE_TYPES.TSImportEqualsDeclaration,
+  AST_NODE_TYPES.TSTypeAliasDeclaration,
+  AST_NODE_TYPES.TSEmptyBodyFunctionExpression,
+];
+
 const baseRule = getESLintCoreRule('semi');
 
 export type Options = util.InferOptionsTypeFromRule<typeof baseRule>;
@@ -38,26 +56,13 @@ export default util.createRule<Options, MessageIds>({
     const checkForSemicolon =
       rules.ExpressionStatement as TSESLint.RuleFunction<TSESTree.Node>;
 
-    /*
-      The following nodes are handled by the member-delimiter-style rule
-      AST_NODE_TYPES.TSCallSignatureDeclaration,
-      AST_NODE_TYPES.TSConstructSignatureDeclaration,
-      AST_NODE_TYPES.TSIndexSignature,
-      AST_NODE_TYPES.TSMethodSignature,
-      AST_NODE_TYPES.TSPropertySignature,
-    */
-    const nodesToCheck = [
-      AST_NODE_TYPES.PropertyDefinition,
-      AST_NODE_TYPES.TSAbstractPropertyDefinition,
-      AST_NODE_TYPES.TSDeclareFunction,
-      AST_NODE_TYPES.TSExportAssignment,
-      AST_NODE_TYPES.TSImportEqualsDeclaration,
-      AST_NODE_TYPES.TSTypeAliasDeclaration,
-      AST_NODE_TYPES.TSEmptyBodyFunctionExpression,
-    ].reduce<TSESLint.RuleListener>((acc, node) => {
-      acc[node as string] = checkForSemicolon;
-      return acc;
-    }, {});
+    const nodesToCheck = SEMI_AST_NODE_TYPES.reduce<TSESLint.RuleListener>(
+      (accumulator, node) => {
+        accumulator[node as string] = checkForSemicolon;
+        return accumulator;
+      },
+      {},
+    );
 
     return {
       ...rules,
